model_abdomenal/project/requirements.txt
torchsummary
imutils
autopep8
sockjs-tornado
pybase64
pycocotools
seaborn
scikit-image==0.17.2
ujson
segmentation-models-pytorch
torchmetrics
libtiff
ttach
albumentations
pydicom
